Refactored for database and handler functions. Added bookingsFunctions and bookingsHandler.
